refactor: ReservationDao를 인터페이스로 추상화하도록 변경
